LayoutTests/imported/w3c/web-platform-tests/css/filter-effects/inheritance-expected.txt


FAIL Property backdrop-filter has initial value none assert_true: backdrop-filter doesn't seem to be supported in the computed style expected true got false
FAIL Property backdrop-filter does not inherit assert_true: expected true got false
FAIL Property color-interpolation-filters has initial value linearrgb assert_equals: expected "linearrgb" but got "linearRGB"
FAIL Property color-interpolation-filters inherits assert_equals: expected "srgb" but got "sRGB"
PASS Property filter has initial value none
PASS Property filter does not inherit
PASS Property flood-color has initial value rgb(0, 0, 0)
PASS Property flood-color does not inherit
PASS Property flood-opacity has initial value 1
PASS Property flood-opacity does not inherit
PASS Property lighting-color has initial value rgb(255, 255, 255)
PASS Property lighting-color does not inherit
